build(Modularized-Backend): Backend is modularized based on partition by feature. Catgeory, product and database are three new features
